pyproject.toml
[tool.poetry]
name = "book-platform"
version = "0.1.0"
description = ""
authors = ["Christoph Andre Gotschlich <christoph.gotschlich@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
httpx = "^0.26.0"
pytest = "^7.4.3"
thefuzz = "^0.20.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
